third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-blocked.html

<!DOCTYPE html>
<html>
    <body>
        <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <p>
        This test passes if no 'FAIL' alert appears.
        </p>
        <script src='fail.js' integrity="ni:///sha256;BpfBw7ivV8q2jLiT13fxDYAe2tJllusRSZ273h2nFSE="></script>
        <script src='fail.js' integrity="ni:///sha384;XVVXBGoYw6AJOh9J/Z8pBDMVVPfkBpngexkA7JqZu8d5GENND6TEIup/tA1v5GPr"></script>
        <script src='fail.js' integrity="ni:///sha512;tbUPioKbVBplr0b1ucnWB57SJWt4x9dOE0Vy2mzCXvH3FepqDZ+07yMK81ytlg0MPaIrPAjcHqba5csorDWtKg=="></script>
    </body>
</html>
